requirements.txt

annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
blis==0.7.11
boto==2.49.0
boto3==1.34.154
botocore==1.34.154
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==7.1.2
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
datasketch==1.6.5
Django==5.0.7
django-ses==4.1.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.6
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
Levenshtein==0.25.1
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
mysqlclient==2.2.4
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.4.0
preshed==3.0.9
profanity==1.1
protobuf==5.27.3
pydantic==1.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
regex==2024.5.15
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.6
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
sympy==1.13.2
textblob==0.18.0.post0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.4
typer==0.3.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==0.10.1
weasel==0.4.1
wikipedia==1.4.0
Wikipedia-API==0.6.0
wrapt==1.16.0
